==> jira/dataset_creation/Jira_dataset/03_combine_dataset.ipynb <==
import pymongo
# Default connection to localhost
myclient = pymongo.MongoClient("mongodb://localhost:27017/")


mydb = myclient["JiraRepos"]
collist = mydb.list_collection_names()
collist

import pandas as pd
import os
# Create dir if not exists
os.makedirs('only_useful_cols', exist_ok=True)
all_dfs = []
for collection in collist:
    try:
        project = pd.read_csv(f'only_useful_cols/{collection}.csv')

        all_dfs.append(project)
        # Make a new column where the rows are the collection name
        project['collection'] = collection
        print(f'Processed {collection}')
   

    except Exception as e:
        print(f'Error  {e} in {collection}')
        continue


if all_dfs:
    df = pd.concat(all_dfs, ignore_index=True)
else:
    print('No CSV files were processed.')
df

# To csv
df.to_csv('all_projects.csv', index=False)
