==> jira/dataset_creation/Jira_dataset/04_merge_priority_dataset.ipynb <==
# Map priority levels into a class for each.
This notebook will map each different priority levels in a common class.
The project with 5 different priority levels are selected.

import pandas as pd
# Import dataset from csv file
df = pd.read_csv(f'all_projects.csv')

# rename columns using rename function
df.rename(columns={'fields.priority.name': 'priority', 'fields.description': 'description', 'fields.project.name': 'project', 'fields.issuetype.name': 'issuetype', 'fields.labels': 'labels'}, inplace=True)
df

# drop priority rows with NaN
df = df.dropna(subset=['priority'])
# rest index
df = df.reset_index(drop=True)
df

# count priority
df['priority'].value_counts().to_frame()[:50]


# Remove issues with no priority level set.
df = df[df['priority'] != 'Unset']
df = df[df['priority'] != 'TBD']
df = df[df['priority'] != 'Undefined']
df = df[df['priority'] != 'Unprioritized']
df = df[df['priority'] != 'Not Evaluated']
df = df[df['priority'] != 'Unknown']


# rest index
df = df.reset_index(drop=True)
df


df['priority'].value_counts().to_frame()[:50]

## Create dataset with projects with 5 different priority levels.
Using definition from Jira docs.
https://support.atlassian.com/jira-service-management-cloud/docs/what-are-priority-levels-in-jira-service-management/

Example:
* 4: Trivial = Lowest
* 3: Minor = Low
* 2: Major  = Medium
* 1: Critical = High
* 0: Blocker = Highest


# Extract test set ofr in the wild
jira = df[df['collection'] == 'Jira']
mongodb = df[df['collection'] == 'MongoDB']
qt = df[df['collection'] == 'Qt']
# Maybe 
jfrog = df[df['collection'] == 'JFrog']

# Remove projects that does not use 5 priority levels
df = df[df['collection'] != 'Jira']
df = df[df['collection'] != 'MongoDB']
df = df[df['collection'] != 'Qt']

df = df[df['collection'] != 'JFrog']
# Project with few issues with priority or bad use of priority levels
df = df[df['collection'] != 'Mojang']
# Remove some other labels that are had to fit into 5 priority levels
# Apache removes rows with priority: Low, Normal, High, Urgent
priorities_to_remove = ['Low', 'Normal', 'High', 'Urgent']
collection_to_match = 'Apache'

# Filter the DataFrame to keep rows that don't match both conditions
df = df[~((df['priority'].isin(priorities_to_remove)) & (df['collection'] == collection_to_match))]
df

# Merge qt, mnogodb and jira back new df
test_set = pd.concat([jira, mongodb, qt, jfrog])
test_set

# To csv
test_set.to_csv('test_sets_projects.csv', index=False)

# Unique collection
test_set['collection'].unique()

test_set['priority'].value_counts().to_frame()[:50]

# show priority in the apache collection
apache = df[df['collection'] == 'Apache']
apache['priority'].value_counts().to_frame()[:50]

# In collection RedHat make priority low == lowest
df.loc[df['collection'].isin(['RedHat']) & (df['priority'] == 'Low'), 'priority'] = 'Lowest'
# Priority in RedHat collection
redhat = df[df['collection'] == 'RedHat']
redhat['priority'].value_counts().to_frame()[:50]

# Unique collection in the training set
df['collection'].unique()

# count priority
df['priority'].value_counts().to_frame()[:50]

# Define mapping for the 'class' column
priority_mapping = {
    # Highest priority
    'Blocker': 'Highest',
    'P0': 'Highest',
    'Urgent': 'Highest',
    'Showstopper': 'Highest',
    'P0: Blocker': 'Highest', # Note: some project uses P0 as the highest priority level,
    'P1-Urgent': 'Highest', # others use P1 as the highest priority level
    'Highest': 'Highest',
    'Blocker - P1': 'Highest',
    # High
    'High': 'High',
    'Critical': 'High',
    'P1': 'High',
    'Severe': 'High',
    'Critical': 'High',
    'P2-High': 'High',
    'P1: Critical': 'High',
    'Critical - P2': 'High',
    # Medium
    'Medium': 'Medium',
    'Major': 'Medium',
    'P2': 'Medium',
    'P3-Medium': 'Medium', 
    'Important': 'Medium',
    'P2: Important': 'Medium',
    'Major - P3': 'Medium',
    # Low
    'Low': 'Low',
    'Minor': 'Low',
    'P3': 'Low',
    'Normal': 'Low',
    'P4-Low': 'Low',
    'P3: Somewhat important': 'Low',
    'Minor - P4': 'Low',
    # Lowest
    'Lowest': 'Lowest',
    'P4': 'Lowest',
    'Trivial': 'Lowest',
    'P5-Trivial': 'Lowest',
    'P4: Low': 'Lowest',
    'Trivial - P5': 'Lowest',
    'P5: Not important': 'Lowest',


}
# Apply mapping
df['class'] = df['priority'].map(priority_mapping)


# Find the priorities that are not mapped to 'Highest' by checking for nulls in 'class'
unmapped_priorities = df[df['class'].isnull()]['priority'].unique()

print("Priorities not mapped to a new class:")
for priority in unmapped_priorities:
    print(f"- {priority}")



# Value counts of class
df['class'].value_counts().to_frame()[:50]

df

# Show count issuetype
df['issuetype'].value_counts().to_frame()[:50]

# Make new df with only issuetype: Technical Debt
technical_debt = df[df['issuetype'] == 'Technical Debt']
# To csv
technical_debt.to_csv('priority_with_TD.csv', index=False)
technical_debt

technical_debt['class'].value_counts().to_frame()[:50]

# Remove issues with Technical Debt issue type
df = df[df['issuetype'] != 'Technical Debt']
df = df.reset_index(drop=True)
df

df["class"].value_counts().to_frame()[:50]

# Save to csv
df.to_csv('all_priority_group_in_classes.csv', index=False)

# Read csv to check if file is saved correctly
df = pd.read_csv('all_priority_group_in_classes.csv')

import os
priority_levels = ['Highest', 'High', 'Medium', 'Low', 'Lowest']

# Saves each class to a separate csv file
for level in priority_levels:
    try:
        # Make dir with level
        os.makedirs(f'{level}', exist_ok=True)
        # df with level class
        df_level = df[df['class'] == level]
        # Save to csv
        df_level.to_csv(f'{level}/{level}.csv', index=False)
        print(f"Saved {level}.csv")
    except Exception as e:
        print(f"An error occurred for level {level}: {str(e)}")


# Read csv to check if file is saved correctly
for level in priority_levels:
    try:
        df = pd.read_csv(f'{level}/{level}.csv')
        print(f"Read {level}.csv")
    except Exception as e:
        print(f"An error occurred while reading {level}.csv: {str(e)}")


## Definition of each priority level according to the atlassian documentation.
* Lowest - Trivial problem with little or no impact on progress. Color: Light grey.
* Low - Minor problem or easily worked around. Color: Dark grey.
* Medium - Has the potential to affect progress. Color: Yellow.
* High - Serious problem that could block progress. Color: Orange.
* Highest - The problem will block progress. Color: A dark red.

#### Compared to github

* High == High and highest 
* medium == medium
* low = low and lowest

# Ranking of priority levels.
Most commonly used priority levels in this dataset.
Using 4-6 levels of priority.
5 levels is most common.
From low to high
## 5 priority levels
### Sonatype, MongoDB, Apache, RedHat, Spring, Sakai(not any trivial), JiraEcosytem, Sonatype
* Trivial
* Minor
* Major
* Critical
* Blocker
### Apache
* P4
* P3
* P2
* P1
* P0

### Hyperledger, Mindville
* Lowest
* Low 
* Medium
* High
* Highest

### RedHat
* Low
* Normal
* Medium
* High
* Urgent

### IntelDOAS:
* P5-Trivial
* P4-Low
* P3-Medium
* P2-High
* P1-Urgent

### SecondLife
* Trivial
* Minor
* Major
* Severe
* Showstopper

### Mojang
* Low
* Normal
* Important
* Critical
* Blocker

## 6 levels
### QT
* P5: Not important (removing this)
* P4: Low
* P3: Somewhat important
* P2: Important
* P1: Critical
* P0: Blocker

### JFrog
* Trivial
* Minor
* Normal
* High
* Critical
* Blocker


## 4 levels
### Apache
* Low
* Normal
* High
* Urgent

### Jira (the org)
* Low
* Medium
* High
* Highest
### Mindville (few issues) can ignore.
* Level 4
* Level 3
* Level 2
* Level 1
