==> jira/dataset_creation/Jira_dataset/05_cleaning_highest_vs_rest.ipynb <==
# Data cleaning
We want the class with labels as numerical value and the body with clean text.

This will remove:
* duplicates
* NaN entires
* non english
* url, html

* make it lowercase
* combine title and body

import pandas as pd
import sys
sys.path.append("../../../scripts_shared/")
from preprocess_text import preprocess_text


file_name = "all_priority_group_in_classes.csv"
df = pd.read_csv(file_name)
df

# Count per priority
df['class'].value_counts()

df['issuetype'].value_counts().to_frame()[:50]

# Unique projects
df['project'].nunique()


# Unique collections
df['collection'].nunique()

# Count per collection
df['collection'].value_counts().to_frame()[:50]

# Drop duplicates by the content of the description
df = df.drop_duplicates(subset=['description'], keep='last')
df.dropna(inplace=True)
df.reset_index(inplace=True)
df.drop(columns=["index"] , inplace= True)
df["class"].value_counts()

print(df["description"][0])

# Convert to string
df["text_str"] = df['description'].astype(str)

# Clean the data.
df["text_clean"] = df["text_str"].map(preprocess_text)

# save df to csv
df.to_csv("all_priority_group_in_classes_clean.csv", index=False)

# Extract only cols need
df = df[["class", "text_clean"]]

# Row with NaN
df[df.isna().any(axis=1)]

# Need to dropna here since cleaning function returns NaN for not english text.
df.dropna(inplace=True)
df.reset_index(drop=True,inplace=True)

df

# Rows with NaN
df[df.isna().any(axis=1)]

null_rows = df[df['text_clean'].isnull()]
null_rows

# Clean dataset with clean text and class.
name = f"clean_{file_name}.csv"
df.to_csv(name, index=False)

import os
priority_levels = ['Highest', 'High', 'Medium', 'Low', 'Lowest']

for level in priority_levels:
    try:
        # Make dir with level
        os.makedirs(f'{level}', exist_ok=True)
        # df with level class
        df_level = df[df['class'] == level]
        # Save to csv
        df_level.to_csv(f'{level}/clean_{level}.csv', index=False)
        print(f"Saved {level}.csv")
    except Exception as e:
        print(f"An error occurred for level {level}: {str(e)}")

# Read csv to check if file is saved correctly
for level in priority_levels:
    try:
        df = pd.read_csv(f'{level}/clean_{level}.csv')
        print(f"Read {level}.csv")
    except Exception as e:
        print(f"An error occurred while reading {level}.csv: {str(e)}")

pri = pd.read_csv(name)
pri

null_rows = pri[pri['text_clean'].isnull()]
null_rows

# Remove rows with NaN
pri = pri.dropna()
# Reset index
pri.reset_index(drop=True, inplace=True)
pri
